Revert "198 Renamed components with suffix "Input" that weren't before"
This reverts commit 01414a0e042db5677c740272325dd252746486f0.
