+++ src/frontend/src/components/Input/GraphInput.tsx
@@ -2,7 +2,7 @@ import { Box } from "@chakra-ui/react";
 import { ModifierInput } from "./ModifierInputComp/ModifierInput";
 import { MiscItemInput } from "./ItemInputComp/MiscItemInput";
 import { BaseInput } from "./ItemBaseTypeInputComp/BaseInput";
-import { IsItemInput } from "./ItemInputComp/IsItemInput";
+import { IsItemInput } from "./ItemInputComp/IsItemProp";
 import { ItemNameInput } from "./ItemInputComp/ItemNameInput";
 import { ItemRarityInput } from "./ItemInputComp/ItemRarityInput";
 import { LeagueInput } from "./LeagueInput";
