+++ src/frontend/src/components/Input/ItemBaseTypeInputComp/CategoryInput.tsx
@@ -4,7 +4,7 @@ import { ItemBaseTypeCategory } from "../../../client";
 import {
   SelectBox,
   SelectBoxOptionValue,
-} from "../StandardLayoutInput/SelectBoxInput";
+} from "../StandardLayoutInput/SelectBox";
 
 interface CategoryInputProps {
   categories: ItemBaseTypeCategory | ItemBaseTypeCategory[];
