+++ src/frontend/src/components/Input/ItemInputComp/MiscItemInput.tsx
@@ -1,5 +1,5 @@
 import { Flex } from "@chakra-ui/layout";
-import { IsItemInput } from "./IsItemInput";
+import { IsItemInput } from "./IsItemProp";
 import { MinMaxInput } from "./MinMaxItemLvlInput";
 import { useEffect, useState } from "react";
 import { Checkbox, CheckboxIcon, Text } from "@chakra-ui/react";
